+++ slice.go
@@ -1,0 +1,58 @@
+package versionary
+
+// Map turns a slice of T1 into a slice of T2 using a mapping function.
+func Map[T1, T2 any](s []T1, f func(T1) T2) []T2 {
+	r := make([]T2, len(s))
+	for i, v := range s {
+		r[i] = f(v)
+	}
+	return r
+}
+
+// Reduce reduces a slice of T1 to a single value using a reduction function.
+func Reduce[T1, T2 any](s []T1, initializer T2, f func(T2, T1) T2) T2 {
+	r := initializer
+	for _, v := range s {
+		r = f(r, v)
+	}
+	return r
+}
+
+// Filter filters values from a slice using a filter function.
+// It returns a new slice with only the elements of s that satisfy the predicate.
+func Filter[T any](s []T, f func(T) bool) []T {
+	var r []T
+	for _, v := range s {
+		if f(v) {
+			r = append(r, v)
+		}
+	}
+	return r
+}
+
+// Contains returns true if the slice contains the provided value.
+func Contains[T comparable](s []T, value T) bool {
+	for _, v := range s {
+		if v == value {
+			return true
+		}
+	}
+	return false
+}
+
+// Batch divides the provided slice of things into batches of the specified maximum size.
+func Batch[T any](items []T, batchSize int) [][]T {
+	batches := make([][]T, 0)
+	batch := make([]T, 0)
+	for _, item := range items {
+		batch = append(batch, item)
+		if len(batch) == batchSize {
+			batches = append(batches, batch)
+			batch = make([]T, 0)
+		}
+	}
+	if len(batch) > 0 {
+		batches = append(batches, batch)
+	}
+	return batches
+}
